Move Django test config into tests
