+++ pyproject.toml
@@ -119,7 +119,8 @@ module = [
 disable_error_code = ["misc"]
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "cross_web.testing._django_settings"
+DJANGO_SETTINGS_MODULE = "testing._django_settings"
+pythonpath = ["tests"]
 markers = [
     "aiohttp: marks tests for aiohttp framework",
     "chalice: marks tests for chalice framework",
